backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5
tensorflow>=2.20.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas==2.2.0
redis==5.0.1
celery==5.3.6
stripe==7.11.0
web3==6.15.0
python-socketio==5.11.0
Pillow==10.2.0
httpx==0.26.0
beautifulsoup4==4.12.3
aiohttp==3.9.1
pyjwt==2.8.0
email-validator==2.1.0
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
transformers==4.37.2
sentence-transformers==2.3.1
pinecone-client==3.0.2
chromadb==0.4.22
